refactor: RacingGame.test 수정
- jest.fn() 제거
- 테스트가 어려운 부분 분리, 가능한 부분만 테스트 작성
