+++ __tests__/__models__/Car.test.js
@@ -1,4 +1,4 @@
-const Car = require('../../src/models/Car');
+const Car = require('../../src/domain/models/Car');
 
 describe('Car', () => {
   test('전진시 위치가 1 증가한다', () => {
