+++ docs/REQUIREMENTS.md
@@ -2,7 +2,6 @@
 
 ### 유닛테스트
 
-- [x] 무작위 값이 4 이상일 경우에만, 전진을 한다.
 - [x] 에러 테스트
   - [x] 출력 메시지 확인
   - [x] 재입력을 받는지 여부 확인
@@ -15,8 +14,6 @@
 - [x] 전진 시도 횟수 유효성 검증
   - [x] 전진 시도 횟수는 숫자만 입력할 수 있다.
   - [x] 전진 시도 횟수는 1회 이상 10회 이하만 입력할 수 있다.
-- [x] 스냅샷이 올바르게 저장된다.
-- [x] 올바른 우승자가 도출되어야 한다.
 - [x] 경주가 끝난 후 프로그램이 올바르게 종료되어야 한다.
 - [x] 모두가 전진하지 못한 경우 예외 처리한다.
 
@@ -42,5 +39,5 @@
 
 - [x] 도메인 로직은 domain/ 하위로, UI 관련 로직은 view/ 하위에서 관리한다.
 - [x] domain/ 하위의 모듈은 view/ 하위의 모듈을 의존하지 않아야 한다.
-- [] 도메인 로직 내에서 테스트하기 어려운 부분을 분리하고, 테스트 가능한 부분에 대해서만 단위 테스트를 작성한다.
-- [] 테스트 코드에서 jest.fn()을 사용하지 않는다.
+- [x] 도메인 로직 내에서 테스트하기 어려운 부분을 분리하고, 테스트 가능한 부분에 대해서만 단위 테스트를 작성한다.
+- [x] 테스트 코드에서 jest.fn()을 사용하지 않는다.
